+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - more documentation
+- set project urls in metadata
 
 ## [v0.15.0] - 2022-09-21